requirements.txt
discogs-client==2.2.2
requests-oauthlib
mp3-tagger==1.0
mutagen==1.47.0
musicbrainzngs>=0.7
oauth2==1.9.0.post1
python-dotenv
pytest==5.4.1
